Extract actions behaviour from Workflow
though actions are related to Workflow
in a way, we can keep those methods in
a seperate file for clarity.
