+++ app/models/workflow.rb
@@ -1,5 +1,4 @@
 require "state_machine"
-require "action"
 
 module Workflow
   class CannotDeletePublishedPublication < RuntimeError; end
@@ -15,7 +14,6 @@ module Workflow
 
     field :state, type: String, default: "draft"
     belongs_to :assigned_to, class_name: "User"
-    embeds_many :actions
 
     state_machine initial: :draft do
       after_transition on: :request_amendments do |edition, transition|
@@ -105,68 +103,12 @@ module Workflow
     text
   end
 
-  def update_user_action(property, statuses)
-    actions.where(:request_type.in => statuses).limit(1).each do |action|
-      # This can be invoked by Panopticon when it updates an artefact and associated
-      # editions. The problem is that Panopticon and Publisher users live in different
-      # collections, but share a model and relationships with eg actions.
-      # Therefore, Panopticon might not find a user for an action.
-      if action.requester
-        set(property, action.requester.name)
-      end
-    end
-  end
-
   def denormalise_users!
     set(:assignee, assigned_to.name) if assigned_to
     update_user_action("creator",   [Action::CREATE, Action::NEW_VERSION])
     update_user_action("publisher", [Action::PUBLISH])
     update_user_action("archiver",  [Action::ARCHIVE])
     self
-  end
-
-  def created_by
-    creation = actions.detect do |a|
-      a.request_type == Action::CREATE || a.request_type == Action::NEW_VERSION
-    end
-    creation.requester if creation
-  end
-
-  def published_by
-    publication = actions.where(request_type: Action::PUBLISH).first
-    publication.requester if publication
-  end
-
-  def archived_by
-    publication = actions.where(request_type: Action::ARCHIVE).first
-    publication.requester if publication
-  end
-
-  def latest_status_action(type = nil)
-    if type
-      self.actions.where(request_type: type).last
-    else
-      most_recent_action(&:status_action?)
-    end
-  end
-
-  def last_fact_checked_at
-    last_fact_check = actions.reverse.find(&:is_fact_check_request?)
-    last_fact_check ? last_fact_check.created_at : NullTimestamp.new
-  end
-
-  def new_action(user, type, options={})
-    actions.create!(options.merge(requester_id: user.id, request_type: type))
-  end
-
-  def new_action_without_validation(user, type, options={})
-    action = actions.build(options.merge(requester_id: user.id, request_type: type))
-    save(validate: false)
-    action
-  end
-
-  def most_recent_action(&blk)
-    self.actions.sort_by(&:created_at).reverse.find(&blk)
   end
 
   def can_destroy?
@@ -187,14 +129,6 @@ module Workflow
 
   def notify_siblings_of_new_edition
     siblings.update_all(sibling_in_progress: self.version_number)
-  end
-
-  def notify_siblings_of_published_edition
-    siblings.update_all(sibling_in_progress: nil)
-  end
-
-  def update_sibling_in_progress(version_number_or_nil)
-    update_attribute(:sibling_in_progress, version_number_or_nil)
   end
 
   def in_progress?
@@ -222,6 +156,26 @@ module Workflow
   end
 
   private
+
+    def notify_siblings_of_published_edition
+      siblings.update_all(sibling_in_progress: nil)
+    end
+
+    def update_sibling_in_progress(version_number_or_nil)
+      update_attribute(:sibling_in_progress, version_number_or_nil)
+    end
+
+    def update_user_action(property, statuses)
+      actions.where(:request_type.in => statuses).limit(1).each do |action|
+        # This can be invoked by Panopticon when it updates an artefact and associated
+        # editions. The problem is that Panopticon and Publisher users live in different
+        # collections, but share a model and relationships with eg actions.
+        # Therefore, Panopticon might not find a user for an action.
+        if action.requester
+          set(property, action.requester.name)
+        end
+      end
+    end
 
     def publish_at_is_in_the_future
       errors.add(:publish_at, "can't be a time in the past") if publish_at.present? && publish_at < Time.zone.now
